requirements.txt
aniso8601==1.3.0
click==6.7
flake8==3.5.0
Flask==0.12.2
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mccabe==0.6.1
nose==1.3.7
pycodestyle==2.3.1
pyflakes==1.6.0
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
SQLAlchemy==1.1.14
Werkzeug==0.12.2
